Refactor dashboard sidebar to take customizations as args
