+++ src/components/user/Dashboard.js
@@ -3,7 +3,15 @@ import { connect } from "react-redux";
 import {hideNavigation, setUserDashboardView} from "../../actions";
 import Calendar from "./schedule/Calendar";
 import Maps from "../Maps";
-import MiniDrawer from "./sidebar/MiniDrawer";
+import InboxIcon from "@material-ui/icons/MoveToInbox";
+import MailIcon from "@material-ui/icons/Mail";
+import DraftsIcon from "@material-ui/icons/Drafts";
+import VideocamIcon from "@material-ui/icons/Videocam";
+import EventIcon from "@material-ui/icons/Event";
+import MapIcon from "@material-ui/icons/Map";
+import DescriptionIcon from "@material-ui/icons/Description";
+import PhotoCameraIcon from "@material-ui/icons/PhotoCamera";
+import {MiniDrawer} from "./sidebar/MiniDrawer";
 
 const viewMap = {
     'Inbox': 'Inbox',
@@ -16,15 +24,29 @@ const viewMap = {
     'Photos': 'Photos'
 }
 
+const iconMap = {
+    'Inbox': <InboxIcon />,
+    'Send Message':  <MailIcon />,
+    'Drafts': <DraftsIcon/>,
+    'E-Visit': <VideocamIcon/>,
+    'Calendar': <EventIcon/>,
+    'Maps': <MapIcon/>,
+    'Visit-Summary': <DescriptionIcon/>,
+    'Photos': <PhotoCameraIcon/>
+}
+
 class UserDashboard extends Component {
     render() {
         this.props.hideNavigation(true);
-        // FIXME: Find a way to make this more responsive, looks bad on small screens
+        // FIXME: Change colors to match home screen
         return (
             <div>
                 <MiniDrawer
                     handleViewChange={ this.props.setUserDashboardView }
                     renderView = { () => viewMap[this.props.view] }
+                    iconMap = { iconMap }
+                    upperItems = { ['Inbox', 'Send Message', 'Drafts'] }
+                    lowerItems = { ['Calendar', 'Maps', 'Photos', 'E-Visit'] }
                 />
             </div>
         )
